feat: 로그인한 사용자만 할일 수정/삭제 가능
사용자가 작성한 할 일이 아닐 시 발생하는 예외 처리

+++ src/main/java/com/example/tododevelop/service/TodoService.java
@@ -62,8 +62,11 @@ public class TodoService {
     }
 
     // 할 일 삭제
-    public void deleteTodo(Long id) {
+    public void deleteTodo(Long id, Long userId) {
         TodoEntity findTodo = findByIdOrElseThrow(id);
+        if (findTodo.getUserEntity().getId() != userId) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "작성자만 삭제 가능합니다.");
+        }
         todoRepository.delete(findTodo);
     }
